Created StarsContainer component
