+++ packages/strapi-admin/admin/src/components/StarsContainer/styles.scss
@@ -1,0 +1,13 @@
+.starsContainer {
+  display: flex;
+  margin-right: 10px;
+  > div {
+    color: #EED348;
+    > i {
+      margin-right: 2px;
+    }
+  }
+  > div:last-child {
+    color: #B3B5B9;
+  }
+}
