+++ packages/strapi-admin/admin/src/translations/en.json
@@ -14,7 +14,7 @@
   "app.components.InstallPluginPage.plugin.support-us.description": "Support us by buying the Strapi t-shirt. It will allow us to keep working on the project and try giving the best possible experience!",
   "app.components.InstallPluginPage.InputSearch.label": " ",
   "app.components.InstallPluginPage.InputSearch.placeholder": "Search a plugin... (ex: authentication)",
-  "app.components.InstallPluginPopup.downloads": "downloads",
+  "app.components.InstallPluginPopup.downloads": "download",
   "app.components.InstallPluginPopup.navLink.description": "Description",
   "app.components.InstallPluginPopup.navLink.screenshots": "Screenshots",
   "app.components.InstallPluginPopup.navLink.avis": "avis",
